pyproject.toml
[tool.poetry]
name = "financebot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.14"
streamlit = "^1.32.2"
langchain-community = "^0.0.30"
langchain-openai = "^0.1.1"
langchain-core = "^0.1.37"
python-dotenv = "^1.0.1"
langchain-experimental = "^0.0.56"
pypdf = "^4.1.0"
faiss-cpu = "^1.8.0"
langchainhub = "^0.1.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

